Add support for EntityZones in ComboZones
Because EntityZones move, they can't just slot into the grid architecture of ComboZones. To fix this, EntityZones are given their own special list in ComboZones that keep track of them separately. Therefore, when you test for a point in a ComboZone, it will check all the non-EntityZones in the correct grid cell, then check all EntityZones. If there are no EntityZones, then that check is skipped and things work exactly as they do now.